Add environment-specific services and update project dependencies
This commit introduces various environment-specific service implementations (Dev, QA, UAT, and Prod). These services are used to return the current environment the application is running in. Additionally, the project's dependencies have been updated in the pom.xml file, including the upgrade of the spring-boot-starter-parent version and the addition of the lombok dependency.